requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
